pyproject.toml
[project]
name = "ai-act-compliance-checker"
version = "0.1.0"
description = "An application that analyzes GitHub repositories against the EU AI Act"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.46.0",
    "langchain>=0.1.0",
    "langgraph>=0.1.0",
    "pygithub>=2.1.1",
    "pypdf>=4.0.0",
    "faiss-cpu>=1.7.4",
    "tiktoken>=0.5.2",
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.3.26",
    "requests>=2.31.0",
    "openai>=1.10.0",
]
